Flexpart.DeterministicInput — Type

DeterministicInput

Object that represents a deterministic input file.

time::Dates.DateTime: Time of the input file

filename::String: Filename of the input file

dirpath::String: Absolute path of the directory

Flexpart.DeterministicOutput — Type

DeterministicInput

Object that represents a deterministic output file.

path::String: Path to the output file

type::String: Type of the output file (ncf, binary)

Flexpart.EnsembleInput — Type

EnsembleInput

Object that represents a ensemble input file.

time::Dates.DateTime: Time of the input file

filename::String: Filename of the input file

member::Int64: Ensemble member number of the input file

dirpath::String: Absolute path of the directory

Flexpart.EnsembleOutput — Type

EnsembleInput

Object that represents a ensemble output file.

path::String: Path to the output file

type::String: Type of the output file (ncf, binary)

member::Int64: Ensemble member number

Flexpart.FlexpartSim — Type

struct FlexpartSim{SimType} <: Flexpart.AbstractFlexDir

FlexpartSim represents the directory where the pathnames file is located. pathnames is the file indicating the paths to the files and directories needed by Flexpart (options, output, input and available).

The FlexpartSim object also indicates the type of the simulation:

FlexpartSim{Deterministic} for deterministic flexpart runs

FlexpartSim{Ensemble} for ensemble flexpart runs

If no type parameter is provided when using FlexpartSim constructors, the default will be Deterministic.

Flexpart.FlexpartSim — Method

FlexpartSim(f::Function) -> Any

Apply the function f to a FlexpartSim created in a temporary directory, and remove all of its content upon completion.

Examples

julia> FlexpartSim() do fpsim default_run(fpsim) end ```

Flexpart.FlexpartSim — Method

FlexpartSim(path::String) -> FlexpartSim{Deterministic}

Read the pathnames file in path to create a FlexpartSim.

Flexpart.FlexpartSim — Method

FlexpartSim() -> FlexpartSim{Deterministic}

Create a FlexpartSim in a temporary directory whith the default options and pathnames. It can be copied afterwards with copy.

The default paths for the pathnames are:

["./options/", "./output/", "./input/", "./AVAILABLE"]

Examples

julia> FlexpartSim()
FlexpartSim{Deterministic} @ /tmp/jl_a8gDYX
pathnames:
        :options => "./options/"
        :output => "./output/"
        :input => "./input/"
        :available => "./AVAILABLE"

Flexpart.FpPathnames — Type

FpPathnames

Object that represents the pathnames file. The paths are expected in the following order:

dirpath

options

output

input

available

Flexpart.copy — Method

copy(fpsim::FlexpartSim, path::String)

Copy an existing FlexpartSim to path.

Flexpart.run — Method

run(fpsim::FlexpartSim{Deterministic}; log) -> Any

Run Flexpart using the paths of fpsim.

Flexpart.save — Method

save(fpsim::FlexpartSim)

Write the current FlexpartSim paths to the pathnames file.

Flexpart.saveabs — Method

saveabs(fpsim::FlexpartSim) -> String

Write the current FlexpartSim paths to the pathnames file. Relative paths are converted to absolute path.

Flexpart.set_cmd_dates! — Method

set_cmd_dates!(
    options::FlexpartOption,
    avs::Available
) -> String

Update the simulation start and stop options from the inputs available in avs.

Flexpart.set_cmd_dates! — Method

set_cmd_dates!(
    options::FlexpartOption,
    start::Dates.DateTime,
    stop::Dates.DateTime
) -> String

Update the simulation start and stop options with start date and stop date.

Flexpart.set_point_release! — Method

set_point_release!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.OptionEntry},
    lon,
    lat
) -> Any

Update a release option release location with lon and lat.

Flexpart.set_release_dates! — Method

set_release_dates!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.OptionEntry},
    start::Dates.DateTime,
    stop::Dates.DateTime
) -> String

Update a release option release start and stop datetime given the start and stop DateTime.

Examples

julia> fpoptions = FlexpartOption(FlexpartSim())
julia> Flexpart.set_release_dates!(fpoptions["RELEASES"][:RELEASE][1], Dates.now(), Dates.now() + Dates.Second(120))

Flexpart.set_release_duration! — Method

set_release_duration!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.OptionEntry},
    start::Dates.DateTime,
    duration::Dates.AbstractTime
) -> String

Update a release option release start and stop datetime given the start and duration.

Examples

julia> fpoptions = FlexpartOption(FlexpartSim())
julia> Flexpart.set_release_duration!(fpoptions["RELEASES"][:RELEASE][1], Dates.now(), Dates.Second(120))

Flexpart.set_specie! — Method

set_specie!(
    fpoptions::FlexpartOption,
    specie::String
) -> Int64

Update the specie number in the RELEASES options from the name specie.

Examples

julia> Flexpart.set_specie!(FlexpartOption(FlexpartSim()), "CH4")
26

Flexpart.specie_number — Method

specie_number(specie::AbstractString) -> Any

Return specie number needed for the RELEASES options from the name specie.

Examples

julia> Flexpart.specie_number("CH4")
26

Flexpart.species_name — Method

species_name() -> Vector

Return the names of the species that are available by default with Flexpart