Flexpart.DeterministicOutput — Type

DeterministicInput

Object that represents a deterministic output file.

path::String

: Path to the output file

type::String

: Type of the output file (ncf, binary)

Flexpart.EnsembleOutput — Type

EnsembleInput

Object that represents a ensemble output file.

path::String

: Path to the output file

type::String

: Type of the output file (ncf, binary)

member::Int64

: Ensemble member number

Flexpart.FlexpartDir — Type

struct FlexpartDir{SimType} <: Flexpart.AbstractFlexDir

FlexpartDir represents the directory where the pathnames file is located. pathnames is the file indicating the paths to the files and directories needed by Flexpart (options, output, input and available).

The FlexpartDir object also indicates the type of the simulation:

FlexpartDir{Deterministic} for deterministic flexpart runs

FlexpartDir{Ensemble} for ensemble flexpart runs

If no type parameter is provided when using FlexpartDir constructors, the default will be Deterministic.

Flexpart.FlexpartDir — Method

FlexpartDir(path::String) -> FlexpartDir{Deterministic}

Read the pathnames file in path to create a FlexpartDir.

Flexpart.FlexpartDir — Method

FlexpartDir(f::Function) -> Any

Apply the function f to a FlexpartDir created in a temporary directory, and remove all of its content upon completion.

Examples

julia> FlexpartDir() do fpdir default_run(fpdir) end ```

Flexpart.FlexpartDir — Method

FlexpartDir() -> FlexpartDir{Deterministic}

Create a FlexpartDir in a temporary directory whith the default options and pathnames. It can be copied afterwards with copy.

The default paths for the pathnames are:

["./options/", "./output/", "./input/", "./AVAILABLE"]

Examples

julia> FlexpartDir()
FlexpartDir{Deterministic} @ /tmp/jl_a8gDYX
pathnames:
        :options => "./options/"
        :output => "./output/"
        :input => "./input/"
        :available => "./AVAILABLE"

Flexpart.FpPathnames — Type

FpPathnames

Object that represents the pathnames file. The paths are expected in the following order:

options

output

input

available

Flexpart.copy — Method

copy(fpdir::FlexpartDir, path::String)

Copy an existing FlexpartDir to path.

Flexpart.run — Method

run(fpdir::FlexpartDir{Deterministic}; log) -> Any

Run Flexpart using the paths of fpdir.

Flexpart.save — Method

save(fpdir::FlexpartDir)

Write the current FlexpartDir paths to the pathnames file.

Flexpart.saveabs — Method

saveabs(fpdir::FlexpartDir)

Write the current FlexpartDir paths to the pathnames file. Relative paths are converted to absolute path.

Flexpart.set_cmd_dates! — Method

set_cmd_dates!(
    options::FlexpartOption,
    avs::Available
) -> String

Update the simulation start and stop options from the inputs available in avs.

Flexpart.set_cmd_dates! — Method

set_cmd_dates!(
    options::FlexpartOption,
    start::Dates.DateTime,
    stop::Dates.DateTime
) -> String

Update the simulation start and stop options with start date and stop date.

Flexpart.set_point_release! — Method

set_point_release!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.FlexpartOptions.OptionEntry},
    lon,
    lat
) -> Any

Update a release option release location with lon and lat.

Flexpart.set_release_dates! — Method

set_release_dates!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.FlexpartOptions.OptionEntry},
    start::Dates.DateTime,
    stop::Dates.DateTime
) -> String

Update a release option release start and stop datetime given the start and stop DateTime.

Examples

julia> fpoptions = FlexpartOption(FlexpartDir())
julia> Flexpart.set_release_dates!(fpoptions["RELEASES"][:RELEASE][1], Dates.now(), Dates.now() + Dates.Second(120))

Flexpart.set_release_duration! — Method

set_release_duration!(
    release::OrderedCollections.OrderedDict{Symbol, Flexpart.FlexpartOptions.OptionEntry},
    start::Dates.DateTime,
    duration::Dates.AbstractTime
) -> String

Update a release option release start and stop datetime given the start and duration.

Examples

julia> fpoptions = FlexpartOption(FlexpartDir())
julia> Flexpart.set_release_duration!(fpoptions["RELEASES"][:RELEASE][1], Dates.now(), Dates.Second(120))

Flexpart.set_specie! — Method

set_specie!(
    fpoptions::FlexpartOption,
    specie::String
) -> Int64

Update the specie number in the RELEASES options from the name specie.

Examples

julia> Flexpart.set_specie!(FlexpartOption(FlexpartDir()), "CH4")
26